# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bolsa de Trabajo",
        "description": "Documentación de la API desarrollada en Laravel para la bolsa de trabajo.",
        "contact": {
            "name": "Soporte",
            "email": "soporte@tuapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor Laravel local"
        }
    ],
    "paths": {
...
        "/api/admin/postulaciones": {
            "get": {
                "tags": [
                    "Postulaciones"
                ],
                "summary": "Listar todas las postulaciones",
                "operationId": "getAllPostulaciones",
                "responses": {
                    "200": {
                        "description": "Lista de postulaciones"
                    }
                }
            },
            "post": {
                "tags": [
                    "Postulaciones"
                ],
                "summary": "Crear una nueva postulación",
                "operationId": "createAplication",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Id_Trabajo",
                                    "Id_Candidato",
                                    "Estado",
                                    "Fecha_Envio"
                                ],
                                "properties": {
                                    "Id_Trabajo": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "Id_Candidato": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "Estado": {
                                        "type": "string",
                                        "example": "pendiente"
                                    },
                                    "Fecha_Envio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-05-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Postulación creada correctamente"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/admin/postulaciones/{id}": {
            "get": {
                "tags": [
                    "Postulaciones"
                ],
                "summary": "Obtener una postulación por ID",
                "operationId": "getPostulacionById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la postulación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Postulación encontrada"
                    },
                    "404": {
                        "description": "Postulación no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Postulaciones"
                ],
                "summary": "Actualizar una postulación",
                "operationId": "updateAplication",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la postulación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Id_Trabajo",
                                    "Id_Candidato",
                                    "Estado",
                                    "Fecha_Envio"
                                ],
                                "properties": {
                                    "Id_Trabajo": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "Id_Candidato": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "Estado": {
                                        "type": "string",
                                        "example": "aceptado"
                                    },
                                    "Fecha_Envio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-05-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Postulación actualizada correctamente"
                    },
                    "404": {
                        "description": "Postulación no encontrada"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Postulaciones"
                ],
                "summary": "Eliminar una postulación",
                "operationId": "deleteAplication",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la postulación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Postulación eliminada correctamente"
                    },
                    "404": {
                        "description": "Postulación no encontrada"
                    }
                }
            }
        },
...
    },
...
}